refactor: 최근 상품 ProductRepository로 병합
