+++ app/src/main/java/woowacourse/shopping/data/datasource/local/ProductLocalDataSource.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.data.datasource.local
 
 import woowacourse.shopping.data.db.RecentProductEntity
 
-interface RecentProductLocalDataSource {
+interface ProductLocalDataSource {
     suspend fun getRecentProducts(limit: Int): Result<List<RecentProductEntity>>
 
     suspend fun getRecentViewedProductCategory(): Result<String?>
